refactor common method to util
