+++ src/app/app-utils.spec.ts
@@ -42,4 +42,11 @@ describe('getEnvironment', () => {
     expect(AppUtils.showNavItems('SomeItems')).toEqual(true);
     expect(AppUtils.showNavItems('accept-terms-and-conditions')).toEqual(false);
   });
+
+  it('isRoleExistsForUser', () => {
+    const mockService = jasmine.createSpyObj('mockService', ['get']);
+    mockService.get.and.returnValue('role1,role2,role3');
+    const roleExists = AppUtils.isRoleExistsForUser('role1', mockService);
+    expect(roleExists).toEqual(true);
+  });
 });
