+++ src/app/guards/allowAcceptTerms.guard.ts
@@ -2,10 +2,11 @@ import { Injectable } from '@angular/core';
 import { CanActivate, Router } from '@angular/router';
 import { of, Observable, Subscription } from 'rxjs';
 import { CookieService } from 'ngx-cookie';
-import * as fromApp from '../../../app/store';
+import * as fromApp from '../store';
 import { Store, select } from '@ngrx/store';
 import {catchError, filter, switchMap, take, tap} from 'rxjs/operators';
 import { TermsAndCondition } from 'src/app/models/TermsAndCondition';
+import { AppUtils } from '../app-utils';
 
 @Injectable({
   providedIn: 'root'
@@ -16,7 +17,7 @@ export class AllowAcceptTermsGuard implements CanActivate {
   }
 
   canActivate(): Observable<boolean> {
-    const isPuiCaseManager = this.isRoleExistsForUser('pui-case-manager', this.cookieService);
+    const isPuiCaseManager = AppUtils.isRoleExistsForUser('pui-case-manager', this.cookieService);
     if (isPuiCaseManager) {
       return this.checkStore(this.store).pipe(
         switchMap(() => of(true)),
@@ -48,10 +49,5 @@ export class AllowAcceptTermsGuard implements CanActivate {
     if (tcConfirmed.hasUserAcceptedTC === 'true') {
       store.dispatch(new fromApp.Go({path: [acceptTcPath]}));
     }
-  }
-
-  isRoleExistsForUser(roleName: string, cookieService: CookieService, cookiename: string = 'roles'): boolean {
-    const userRoles = cookieService.get(cookiename);
-    return userRoles && userRoles.indexOf(roleName) >= 0;
   }
 }
